ditch ceperate c# class since i can't access mono methods
