--- Assets/Scripts/UI/TabbedMenuController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 61224502b595a9848a13d8867fa52824
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
