+++ Assets/UI Toolkit/TabbedMenu/TabbedMenu.uxml
@@ -12,7 +12,9 @@
         <ui:VisualElement name="MapContent" class="mapBackground unselectedContent" style="flex-direction: row; height: 100%; width: 100%; flex-wrap: nowrap;">
             <ui:VisualElement name="MapContainer" class="mapContainer">
                 <ui:VisualElement name="MapHolder" class="mapHolder">
-                    <ui:VisualElement name="Map" class="mapImage" style="flex-grow: 1; background-image: url(&quot;project://database/Assets/UI%20Toolkit/TabbedMenu/Map.png?fileID=2800000&amp;guid=db9889df8c131da43b36ad102c0f905e&amp;type=3#Map&quot;);" />
+                    <ui:VisualElement name="Map" class="mapImage" style="flex-grow: 1; background-image: url(&quot;project://database/Assets/UI%20Toolkit/TabbedMenu/Map.png?fileID=2800000&amp;guid=db9889df8c131da43b36ad102c0f905e&amp;type=3#Map&quot;);">
+                        <ui:VisualElement name="Player" style="flex-grow: 0; width: 20px; height: 20px; background-color: rgb(0, 0, 0); translate: 0 0; transform-origin: left top;" />
+                    </ui:VisualElement>
                 </ui:VisualElement>
             </ui:VisualElement>
             <ui:VisualElement name="ClockContainer" class="clockContainer">
